Function roadrunner::request_for_response

pub fn request_for_response(request: Request,
                            core: &mut Core)
                            -> Result<Response, Error>

This function provides a low level API interface that one can use if the high level API is not sufficient.

RestClient and RestClientMethods ony provide limited methods to customize a request. This function, on the other hand, gives users the complete control of how a request is configured.